Added basic UI for home page
Created a basic UI for the home page. Made changes to style.css and home.ejs the most. I also added J.A.D.A to the titles of the login and home pages.

I also have some extra json files apparently? Not sure about those ones if I'm honest.